Change scheduler factory to care about loopers, not handlers.
A Looper is the currency of Android-specific work scheduling that we should be exposing. Handler is only a helper through which we access the Looper.
