+++ rxandroid/src/main/java/rx/android/schedulers/LooperScheduler.java
@@ -1,0 +1,131 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package rx.android.schedulers;
+
+import android.os.Handler;
+import android.os.Looper;
+import android.os.Message;
+import java.util.concurrent.TimeUnit;
+import rx.Scheduler;
+import rx.Subscription;
+import rx.android.plugins.RxAndroidPlugins;
+import rx.android.plugins.RxAndroidSchedulersHook;
+import rx.exceptions.OnErrorNotImplementedException;
+import rx.functions.Action0;
+import rx.plugins.RxJavaPlugins;
+import rx.subscriptions.Subscriptions;
+
+class LooperScheduler extends Scheduler {
+    private final Handler handler;
+
+    LooperScheduler(Looper looper) {
+        handler = new Handler(looper);
+    }
+
+    LooperScheduler(Handler handler) {
+        this.handler = handler;
+    }
+
+    @Override
+    public Worker createWorker() {
+        return new HandlerWorker(handler);
+    }
+
+    static class HandlerWorker extends Worker {
+        private final Handler handler;
+        private final RxAndroidSchedulersHook hook;
+        private volatile boolean unsubscribed;
+
+        HandlerWorker(Handler handler) {
+            this.handler = handler;
+            this.hook = RxAndroidPlugins.getInstance().getSchedulersHook();
+        }
+
+        @Override
+        public void unsubscribe() {
+            unsubscribed = true;
+            handler.removeCallbacksAndMessages(this /* token */);
+        }
+
+        @Override
+        public boolean isUnsubscribed() {
+            return unsubscribed;
+        }
+
+        @Override
+        public Subscription schedule(Action0 action, long delayTime, TimeUnit unit) {
+            if (unsubscribed) {
+                return Subscriptions.unsubscribed();
+            }
+
+            action = hook.onSchedule(action);
+
+            ScheduledAction scheduledAction = new ScheduledAction(action, handler);
+
+            Message message = Message.obtain(handler, scheduledAction);
+            message.obj = this; // Used as token for unsubscription operation.
+
+            handler.sendMessageDelayed(message, unit.toMillis(delayTime));
+
+            if (unsubscribed) {
+                handler.removeCallbacks(scheduledAction);
+                return Subscriptions.unsubscribed();
+            }
+
+            return scheduledAction;
+        }
+
+        @Override
+        public Subscription schedule(final Action0 action) {
+            return schedule(action, 0, TimeUnit.MILLISECONDS);
+        }
+    }
+
+    static final class ScheduledAction implements Runnable, Subscription {
+        private final Action0 action;
+        private final Handler handler;
+        private volatile boolean unsubscribed;
+
+        ScheduledAction(Action0 action, Handler handler) {
+            this.action = action;
+            this.handler = handler;
+        }
+
+        @Override public void run() {
+            try {
+                action.call();
+            } catch (Throwable e) {
+                // nothing to do but print a System error as this is fatal and there is nowhere else to throw this
+                IllegalStateException ie;
+                if (e instanceof OnErrorNotImplementedException) {
+                    ie = new IllegalStateException("Exception thrown on Scheduler.Worker thread. Add `onError` handling.", e);
+                } else {
+                    ie = new IllegalStateException("Fatal Exception thrown on Scheduler.Worker thread.", e);
+                }
+                RxJavaPlugins.getInstance().getErrorHandler().handleError(ie);
+                Thread thread = Thread.currentThread();
+                thread.getUncaughtExceptionHandler().uncaughtException(thread, ie);
+            }
+        }
+
+        @Override public void unsubscribe() {
+            unsubscribed = true;
+            handler.removeCallbacks(this);
+        }
+
+        @Override public boolean isUnsubscribed() {
+            return unsubscribed;
+        }
+    }
+}
